refactor(api): transform sirene stock port into a generic data gouv port
