+++ packages/api/src/adapters/outputs/api/data-gouv/data-gouv.port.ts
@@ -1,0 +1,3 @@
+export interface DataGouvPort {
+    getFile(): Promise<unknown>;
+}
