--- packages/api/src/adapters/outputs/api/sirene/sirene-stock-unite-legale.port.ts
@@ -1,3 +1,0 @@
-export interface SireneStockUniteLegalePort {
-    getZip(): Promise<unknown>;
-}
